Break into manageable components
